feat: refactor data source configuration to support dynamic environment variables
fix: update migration command for orders and services applications

feat: enhance appointment module with transaction and notification services

feat: implement transaction service with Paystack integration and wallet management

feat: add email DTOs for appointment and payment notifications

feat: introduce wallet entity and related functionality in transactions

chore: add Paystack configuration to the application

refactor: update transaction entity to include provider_id and enhance transaction handling

chore: update patterns for transactions and notifications to include new events

+++ apps/orders/src/entities/appointment.entity.ts
@@ -27,7 +27,7 @@ export class Appointment extends BaseClass {
   @Column({ type: 'text', nullable: true })
   join_link?: string;
 
-  @Column({ type: 'int', nullable: true })
+  @Column({ type: 'bigint', nullable: true })
   meeting_id?: number;
 
   @Column({ type: 'decimal', precision: 10, scale: 2, nullable: true })
